requirements.txt
absl-py==0.15.0
astunparse==1.6.3
attrdict==2.0.1
babel==2.10.3
bce-python-sdk==0.8.74
boto3==1.24.57
botocore==1.27.57
cached-property==1.5.2
cachetools==5.2.0
click==8.1.3
cssselect==1.1.0
cssutils==2.5.1
cycler==0.11.0
cython==0.29.32
easydict==1.9
et-xmlfile==1.1.0
flask==2.2.2
flask-babel==2.0.0
flask-jwt-extended==4.4.4
flatbuffers==1.12
fonttools==4.36.0
future==0.18.2
gast==0.4.0
google-auth==2.10.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
imageio==2.21.1
imgaug==0.4.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
jinja2==3.1.2
jmespath==1.0.1
keras-nightly==2.5.0.dev2021032900
keras-preprocessing==1.1.2
kiwisolver==1.4.4
lmdb==1.3.0
lxml==4.9.1
markdown==3.4.1
markupsafe==2.1.1
matplotlib==3.5.3
networkx==2.6.3
numpy==1.19.5
oauthlib==3.2.0
opencv-contrib-python==4.4.0.46
opencv-python==4.6.0.66
openpyxl==3.0.10
packaging==21.3
paddleocr==2.5.0.3
pandas==1.3.5
premailer==3.10.0
protobuf==3.19.4
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyclipper==1.3.0.post3
pycryptodome==3.15.0
pyjwt==2.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
pywavelets==1.3.0
pyyaml==6.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.0
scikit-image==0.19.3
scipy==1.7.3
shapely==1.8.4
six==1.15.0
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
tifffile==2021.11.2
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.64.0
typing-extensions==3.7.4.3
visualdl==2.3.0
werkzeug==2.2.2
wrapt==1.12.1
zipp==3.8.1